+++ src/components/organisms/proposal-review.tsx
@@ -5,12 +5,16 @@ import type {
   ProposalFormValues,
   ProposalSendFormValues,
   ProposalStakeFormValues,
+  ProposalUpdateGroupFormValues,
 } from 'types'
 import { SPACING } from 'util/constants'
 
 import { Heading, Stack, Text } from '@/atoms'
 import { FormCard } from '@/molecules/form-card'
 import { FormSubmitHiddenButton } from '@/molecules/form-footer'
+import { SendReview } from '@/molecules/proposal-send-review'
+import { StakeReview } from '@/molecules/proposal-stake-review'
+import { UpdateGroupReview } from '@/molecules/proposal-update-group-review'
 import { ReviewItem } from '@/molecules/review-item'
 import { Truncate } from '@/molecules/truncate'
 
@@ -70,71 +74,9 @@ function renderAction(action: ProposalAction, groupPolicyAddress: string) {
           values={action.values as ProposalStakeFormValues}
         />
       )
+    case 'update-group':
+      return <UpdateGroupReview values={action.values as ProposalUpdateGroupFormValues} />
     default:
       return null
   }
-}
-
-export const SendReview = ({
-  groupPolicyAddress,
-  values,
-}: {
-  groupPolicyAddress: string
-  values: ProposalSendFormValues
-}) => {
-  return (
-    <FormCard title="Send">
-      <Stack spacing={SPACING.formStack}>
-        <ReviewItem label="Type">{values.sendType}</ReviewItem>
-        <ReviewItem label="From Address">
-          <Truncate
-            clickToCopy
-            headLength={18}
-            tailLength={18}
-            text={groupPolicyAddress}
-          />
-        </ReviewItem>
-        {'toAddress' in values && (
-          <ReviewItem label="To Address">{values.toAddress}</ReviewItem>
-        )}
-        <ReviewItem label="Amount">{`${values.amount} ${values.denom}`}</ReviewItem>
-      </Stack>
-    </FormCard>
-  )
-}
-
-export const StakeReview = ({
-  groupPolicyAddress,
-  values,
-}: {
-  groupPolicyAddress: string
-  values: ProposalStakeFormValues
-}) => {
-  return (
-    <FormCard title="Stake">
-      <Stack spacing={SPACING.formStack}>
-        <ReviewItem label="Type">{values.stakeType}</ReviewItem>
-        <ReviewItem label="Delegator">
-          <Truncate
-            clickToCopy
-            headLength={18}
-            tailLength={18}
-            text={groupPolicyAddress}
-          />
-        </ReviewItem>
-        {'validator' in values && (
-          <ReviewItem label="Validator">{values.validator}</ReviewItem>
-        )}
-        {'fromValidator' in values && (
-          <ReviewItem label="From validator">{values.fromValidator}</ReviewItem>
-        )}
-        {'toValidator' in values && (
-          <ReviewItem label="To validator">{values.toValidator}</ReviewItem>
-        )}
-        {'amount' in values && (
-          <ReviewItem label="Amount">{`${values.amount} ${values.denom}`}</ReviewItem>
-        )}
-      </Stack>
-    </FormCard>
-  )
 }
